### 2018-04-24 #212 @shaunpwilkinson
Submission: phylogram
The phylogram R package is a tool for for developing  phylogenetic trees as deeply-nested lists known as "dendrogram" objects. It provides functions for importing and exporting trees in  parenthetic text format, as well as several tools for command-line tree manipulation and conversion between "dendrogram" and "phylo" class objects.

```
Package: phylogram
Title: Dendrograms for Evolutionary Analysis
Version: 2.0.1.9000
Authors@R: person("Shaun", "Wilkinson", role = c("aut", "cre"),
                     email = "shaunpwilkinson@gmail.com")
Author: Shaun Wilkinson [aut, cre]
Maintainer: Shaun Wilkinson <shaunpwilkinson@gmail.com>
Description: Contains functions for importing and exporting 'dendrogram'
    objects in parenthetic text format, as well as several
    functions for command-line tree manipulation.
    R (>= 3.0.0),
    stats
    ape (>= 4.0),
```

https://github.com/shaunpwilkinson/phylogram

[e.g., "data extraction, because the package parses a scientific data file format"]

Data extraction, the package contains a text parser for trees in the Newick parenthetic text format, and functions for converting between dendrogram and phylo object types.

Biologists and bioinformaticists wishing to access the many analytical and tree-visualization functions available for dendrogram objects in R.

There are other packages that can parse Newick strings to create phylo objects (e.g. ape, ggtree) but none that we are aware of that enable two-way conversion between phylo and dendrogram objects.

    - [x] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.1227670

No errors or warnings.

Tal Galili (talgalili)
Zuguang Gu (jokergoo)
Michael Krabbe Borregaard (mkborregaard)
Guangchuang Yu (GuangchuangYu)
