### 2018-05-01 #212 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @shaunpwilkinson   !!

- You can put a ropensci review badge in your README

- Goodpractice output for you and reviewers to consider, for reviewers to consider:

```r
── GP phylogram ───────────

It is good practice to

    R/ladder.R:33:NA
    R/ladder.R:34:NA
    R/ladder.R:35:NA
    R/ladder.R:36:NA
    R/ladder.R:37:NA
    ... and 63 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with other packages. Use "Imports" instead.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/prune.R:66:1
    R/prune.R:96:1
    R/read.R:74:1
    R/read.R:151:1
    R/write.R:32:1
    ... and 2 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R/ladder.R:35:21
    R/prune.R:43:21
    R/read.R:113:25
    R/read.R:114:26
    R/read.R:158:21
    ... and 3 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.
  ✖ not use exportPattern in NAMESPACE. It can lead to exporting functions unintendedly. Instead, export functions that constitute the external API of
    your package.
```

Seeking reviewers now 🕐

---

Reviewers: xx xx
Due date: xx
