### 2018-05-14 #212 @wcornwell
## Package Review

-   \[x\] As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   \[x\] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
-   \[x\] **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   \[x\] **Vignette(s)** demonstrating major functionality that runs successfully locally
-   \[x\] **Function Documentation:** for all exported functions in R help
-   \[x\] **Examples** for all exported functions in R Help that run successfully locally
-   \[x\] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   \[x\] **Installation:** Installation succeeds as documented.
-   \[x\] **Functionality:** Any functional claims of the software been confirmed.
-   \[ \] **Performance:** Any performance claims of the software been confirmed.
-   \[ \] **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   \[x\] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

-   \[ \] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 3 ---

### Review Comments

Thanks for this package. I didn't know anything really about the dendrogram class in R before looking into this. And it does look like a useful thing for the R-phylo-universe to have ways to go back and forth between the edge matrix style and the nested list style representation. There might be interesting applications in between taxonomy and phylogenies, and it's always good to have a full suite of tools to efficiently translate between data structures.  As taxonomy moves to Non-linnaean data sets, the nested list structure might become increasingly useful.

I thought I would get to this quickly, and just do some of my normal research with this new tool and see how it goes. The tree that we have been using for the last several years is available (here)\[<https://datadryad.org//resource/doi:10.5061/dryad.63q27>\]. It's a ~30,000 tip tree for angiosperms, and typically we cut it down to a subset for whatever the task at hand is, but when I tried to load it using `read.dendrogram`, it still wasn't finished loading after 10 minutes. (The tree loads in less than a second using `ape::read.tree`). So then I did a little investigation. First I checked the size of the `dendrogram` objects using this code:

``` r
evolve.tree <- function(size){
  object.size(diversitree::tree.bd(pars=c(1,0),max.taxa=size))
}

evolve.dendro <- function(size){
  object.size(phylogram::as.dendrogram(diversitree::tree.bd(pars=c(1,0),max.taxa=size)))
}

x <- seq(2, 200, 10)
df <- data.frame(ape.size=purrr::map_dbl(x,evolve.tree),
                 dendro.size=purrr::map_dbl(x,evolve.dendro))

library(ggplot2)
ggplot(df, aes(x=ape.size,y=dendro.size)) + geom_point()+theme_bw()
```
It turns out `dendrogram` objects are about 5 times as big as `phylo` objects, which isn't a big deal.

Then I checked read in speeds:

``` r
read_tree_dendro<-function(size){
  ape::write.tree(diversitree::tree.bd(pars=c(1,0),max.taxa=size),"twoh.tre")
  system.time({phylogram::read.dendrogram("twoh.tre")})[3]
}

read_tree_ape <- function(size){
  ape::write.tree(diversitree::tree.bd(pars=c(1,0),max.taxa=size),"twoh.tre")
  system.time({ape::read.tree("twoh.tre")})[3]
}

set.seed(42)
x <- seq(100,3000,150)
df <- data.frame(phylo_size=x,
               ape.speed=purrr::map_dbl(x, read_tree_ape),
               dendro.speed=purrr::map_dbl(x, read_tree_dendro))

ggplot(df, aes(x=ape.speed,y=dendro.speed,col=phylo_size)) + geom_point() + theme_bw()
```

``` r
ggplot(df, aes(x=phylo_size,y=dendro.speed,col=ape.speed)) + geom_point() + theme_bw()
```

So I guess there is something non-linear about the performance of `read.dendrogram` relative to both `ape::read.tree` and tree size. `ape::read.tree` seems to be linear and about two orders of magnitude faster (although with some mysterious variation sometimes).

I tried to read the `read.dendrogram` code, but I couldn't really intuit where the performance issues are. There are some magic numbers in the code, which might be one place to start. Also I have trouble reading the nested functions, but that might be because of my original training in C--I know this style is common in many languages. But for the big picture, it's worth figuring out the best way to optimise from a higher level. I afraid I'm not very knowledgable about about how R allocates memory to create deeply nested lists, but it's very likely that there are people around `Ropensci` who are, and who may be willing to give constructive advise on optimisation?

Phylogenies are getting bigger all the time: there was a 300,000 tip tree for plants published this week. And there are some giant bacterial trees out also. So for the functionality of this package, designing and optimising the code for at least medium sized trees is pretty crucial.

### other notes:

-   I'm sure there are some use cases for comparative methods where a nested list will be more useful than an edge matrix. Be nice to have a worked example in the vignette. I would also be worth comparing the performance of `ape::drop.tip` with `phylogram::prune`.

-   Reading in problematic newick strings is kind of a nightmare because there are so many flavors of errors, but there are probably some cases where `read.dendrogram` returns a warning, when it really should give an informative error message rather than the current behavior which is to return a warning and a garbled tree.

``` r
library(phylogram)
newick <- "(A:0.1,B:0.2,C:0.3),D:0.4:0.5);"
x <- read.dendrogram(text = newick)
```

    ## Warning in read.dendrogram(text = newick): Possibly incomplete tree parse

``` r
x
```

    ## 'dendrogram' leaf 'D:0.4', at height 0

and also

``` r
newick <- "(A:0.1,(B:0.2,)C:0.3,D:0.4);"
x <- read.dendrogram(text = newick)
```

    ## Warning in read.dendrogram(text = newick): Possibly incomplete tree parse

``` r
x
```
