### 2018-05-14 #212 @wcornwell
#### figures from above

![unnamed-chunk-19-1](https://user-images.githubusercontent.com/5534161/39980639-e7349666-5787-11e8-8ee9-bc62255cdec1.png)

![unnamed-chunk-20-1](https://user-images.githubusercontent.com/5534161/39980645-ee0356ee-5787-11e8-8ebc-1e15a2715e57.png)

![unnamed-chunk-20-2](https://user-images.githubusercontent.com/5534161/39980652-f109825a-5787-11e8-9329-cd8a89f92ea4.png)
