### 2018-05-18 #212 @shaunpwilkinson
Thank you very much for your review @wcornwell I'll make it a priority to optimize the parser to better handle those bigger trees and add more informative error messages for when the tree parse fails. Will aim to get these and your other suggestions seen to in the next week or two. Thanks again for the feedback!
