### 2018-05-23 #212 @shaunpwilkinson
Hi @sckott, I have now made the changes suggested by @wcornwell, including a major overhaul of the read.dendrogram and as.dendrogram.phylo functions which are now much more efficient. Just wondering am I okay to push the changes to GitHub or would you prefer I leave the repo alone until the other review comes in?
