### 2018-05-24 #212 @sckott
@shaunpwilkinson great, thanks. Can you please include a response here to @wcornwell suggestions, it doesn't have to be exactly point by point, but a summary at least.

We don't really have rules about how changes are made with respect to the two different reviewers, but I think it's a good idea to wait for the 2nd review to come in.
