### 2018-05-29 #212 @BenJWard
---

This package will certainly be useful to convert the specific tree classes in
the "ape" family of packages of Paradis et al. into the R dendrogram class and back,
especially since R and other packages contain functionality that works on trees,
such as general hierarchical clustering and classification techniques, that both
statisticians and biologists alike are interested in.

#### Checking, installation, and testing

:+1: `devtools::check`, `devtools::spell_check`, and `devtools::test` all return
with no notes, warnings, errors, or obvious typographical errors.
I do not find any errors that may have been missed on the author's system.

The coverage of the source code is good:

```sh
phylogram Coverage: 85.18%
R/read.R: 76.26%
R/prune.R: 79.73%
R/write.R: 94.74%
R/utils.R: 94.87%
R/asdendrogram.R: 100.00%
R/asphylo.R: 100.00%
R/ladder.R: 100.00%
```

##### Suggestion
Coverage could be improved slightly in read.R and prune.R.

In particular, it would be good to add to the test suite some cases testing that
errors are indeed thrown, when they are expected to be thrown - buy giving
package functions deliberately incorrect input, and using `testthat` to catch
the errors with `expect_error`. This would be good to ensure the Newick reading
is operating as expected and rejecting input when it should, as Newick can be
a problematic format due to its flexibility.

#### Useage

Going through the examples in the vignettes was simple and all examples ran for me.

I won't comment on performance as that has already been addressed in previous
posts. All my examples and files are also not so large that performance is an issue.
In my experience with different tree representations, a nested data structure such as a dendrogram is less efficient for some operations cf. to a contiguous block of data due to spatial locality. The nested data structures also generally require more space.
However, a nested structure is faster and easier to code operations involving
structural edits - prunes, re-grafts and other rearrangements. So it is great
that this package will allow the user to choose the desired tree representation based on need.

##### Suggestions

In the vignette, I think following the instructions on manipulating trees directly i.e.:

```
x <- list(1, list(2, 3))
## attach "leaf" and "label" attributes to leaf nodes
attr(x[[1]], "leaf") <- TRUE
attr(x[[2]][[1]], "leaf") <- attr(x[[2]][[2]], "leaf") <- TRUE
attr(x[[1]], "label") <- "A"
attr(x[[2]][[1]], "label")
..... and so on ....
```

Is important to demonstrate how the trees can be built up from a vanilla R linked
data structure, but this style becomes difficult to read quite quickly for non-trivial
trees.

I think that immediately following this, the vignette could demonstrate how to
build the same tree, but by using functions that define a set of verbs
common to phylogentics (like `prune`, `graft`, `drop.tips` etc.), on a set of nodes.

I think that if this was added it would make users much more comfortable with manipulating trees in a much more semantic way.

In the future - and this is not limited to this package, it would be cool to see
functions such as `ladderize` (`ape`) and `ladder` (`phylogram`) standardized to a
consistent name, perhaps even defined as one function, but with multiple class methods.

#### Concluding

This is a nice and simple R package, which adds compatibility between a few different
classes from a set of related, but sometimes difficult to connect packages.
My suggestions are minor improvements to the package that I think would improve the submission, but it is already very useable.

Thank you for this package :smile:
