### 2018-05-30 #212 @shaunpwilkinson
Thanks so much for your review @BenJWard . @sckott I'm in the process of incorporating the edits and updating the vignette now - and should have the response done either today or tomorrow. Thanks again all!
