### 2018-05-31 #212 @shaunpwilkinson
Hi @sckott, I have now pushed the changes to GitHub and will release a new version of the package on CRAN shortly. We are pleased that the reviewers found the package useful, and are very happy with the improvements made to the package and vignette. Our responses to the reviewer comments/suggestions are listed below.

---

@wcornwell and @BenJWard thank you both so much for the extremely helpful reviews and advice on improving the package. The detailed write-ups you provided were an immense help for optimizing the package functions and making vignette more user friendly. We carried out a fairly major overhaul on the package (see commit # `5b69d36`), that involved writing fast(er) new functions for converting between dendrogram and phylo objects, and getting rid of the old Newick text parser that suffered from some major performance issues and problems with the non-standard parenthetic formatting semantics. We also added more unit tests and generally improved the error management throughout the package. Point by point responses to each of the specific suggestions are listed below. Once again, that you both very much for all of your effort on this so far. We are delighted with the improvements to the package, and hope that the community will find it useful.

Review 1 - @wcornwell

Comment: There is something non-linear about the performance of `read.dendrogram` relative to both `ape::read.tree` and tree size. `ape::read.tree` seems to be linear and about two orders of magnitude faster … for the functionality of this package, designing and optimising the code for at least medium sized trees is pretty crucial.

Response: The previous version of `read.dendrogram` used regular expression pattern matching to read in Newick strings and recursively create nested-list dendrogram objects. In hindsight this was was always going to be much slower than ape’s well-optimized `read.tree`, which calls compiled C code to read in the strings. So we reluctantly decided to jettison the parser, focusing instead on optimizing the `as.dendrogram.phylo` function, which `read.dendrogram` now wraps to convert Newick strings into dendrograms via an intermediate “phylo” object. This resulted in a significant speedup in `read.dendrogram` - e.g. reading in the linked vascular plant tree in around ~ 6 seconds. This is still fairly slow in comparison to `ape::read.tree`, and we will work on optimizing it further in the future, but I guess we will always be limited with dendrogram objects in terms of speed and memory efficiency, particularly in comparison to matrix-based trees. Nevertheless, we think this overhaul has improved the package immensely, and we appreciate both reviewers wisely pointing it out as a speed bottleneck most in need of attention.

Comment: I'm sure there are some use cases for comparative methods where a nested list will be more useful than an edge matrix. Be nice to have a worked example in the vignette.

Response: Great idea, we have now included an example in the vignette on how to convert between phylo and dendrogram objects to create tanglegrams using the **dendextend** package. This seems to be a commonly encountered issue (see threads [here](https://bioinformatics.stackexchange.com/questions/923/how-to-convert-a-phylogeny-to-a-dendrogram-in-r), [here](https://stackoverflow.com/questions/7445684/how-to-convert-a-tree-to-a-dendrogram-in-r) ,  and [here](https://www.biostars.org/p/4576/) ), so I think this constitutes a really important addition to the vignette.

Comment: It would also be worth comparing the performance of `ape::drop.tip` with `phylogram::prune`.

Response: Yes we agree, and have added a couple of sentences to the vignette as follows:  Leaf nodes and internal branching nodes can be removed using the function `prune`, which identifies and recursively deletes nodes based on pattern matching of “label” attributes. This is slower than the ape function `drop.tip`, but offers the benefits of versatile string matching using regular expressions, and the ability to remove inner nodes (and by extension all of their subnodes) that feature matching “label” attributes.

Comment: Reading in problematic newick strings is kind of a nightmare because there are so many flavors of errors, but there are probably some cases where read.dendrogram returns a warning, when it really should give an informative error message rather than the current behavior which is to return a warning and a garbled tree.

Response: This was another factor in our decision to decommission the parser and instead focus on extending ape’s (much better) `read.tree` function to enable dendrogram capability.  Given the amount of work Paradis and others have put into `ape::read.tree` it seemed a bit like reinventing the wheel developing and maintaining a separate Newick parser for dendrograms.

Review 2-  @BenJWard

Comment: Coverage could be improved slightly in read.R and prune.R. In particular, it would be good to add to the test suite some cases testing that errors are indeed thrown, when they are expected to be thrown - by giving package functions deliberately incorrect input, and using testthat to catch the errors with `expect_error`. This would be good to ensure the Newick reading is operating as expected and rejecting input when it should, as Newick can be a problematic format due to its flexibility.

Response: We have now expanded the unit tests to cover 95% of the code (including 100% for read.R and 96% for prune.R) and have also added an `expect_error` for multiPhylo objects. We expect that any other text-formatting issues will now be picked up by `ape::read.tree`.

Comment: In the vignette, I think following the instructions on manipulating trees … is important to demonstrate how the trees can be built up from a vanilla R linked data structure, but this style becomes difficult to read quite quickly for non-trivial trees. I think that immediately following this, the vignette could demonstrate how to build the same tree, but by using functions that define a set of verbs common to phylogentics (like `prune`, `graft`, `drop.tips` etc.), on a set of nodes. I think that if this was added it would make users much more comfortable with manipulating trees in a much more semantic way.

Response: This is a really good point, and would certainly help users get started with some basic manipulation of dendrogram objects. We have now rearranged the vignette to demonstrate the ground-up tree building towards the end (after the basic reading, writing and conversion functions) and added a new example showing how to build and manipulate the same basic tree object with the package functions. We think this adds a valuable contribution to the tutorial.

Comment: In the future - and this is not limited to this package, it would be cool to see
functions such as `ladderize` (ape) and `ladder` (phylogram) standardized to a consistent name, perhaps even defined as one function, but with multiple class methods.

Response: Yes we agree, and did consider adding a new method to ladderize for dendrogram objects, but unfortunately there was a clash with one of the functions in the dendextend package and so we decided to keep them separate. Of course we are happy to take advice on this!
