### 2018-06-08 #212 @sckott
@wcornwell are you happy with the changes made?
