### 2018-06-11 #212 @sckott
@shaunpwilkinson both reviewers are on board with changes. I'll do a quick peek ...
