### 2018-06-11 #212 @sckott
@shaunpwilkinson  one question:

- looks like your `NAMESPACE` file is created by hand. is that correct?  any reason not to create it using `devtools::document()` or similar? Seems less error prone that way, but maybe you have a good reason?
