### 2018-06-11 #212 @shaunpwilkinson
Hi @sckott I spent ages trying to get the NAMESPACE right using roxygen tags and `devtools::document` but gave up in the end, since I couldn't figure out how to get it to export methods for the `as.dendrogram` and `as.phylo` generics correctly. Also I remember seeing a comment from Dirk Eddelbuettel a while back recommending against automatic generation of the NAMESPACE, so I figured I'd just write it manually. I'm happy to change it if you prefer but might need some help with the roxygen tags!
