### 2018-06-12 #212 @shaunpwilkinson
That's great news - thanks so much @sckott !

I'll get to those tasks today.

Quick question - I'm planning to submit the paper to JOSS and wondering if it will be a problem if the `paper.md` file and package vignette are the same? I figured it didn't make much sense to write a whole separate vignette, since the paper includes all of the background and examples necessary to get started with the package. But I'm aware that some journals won't consider a manuscript if it is similar to a package vignette that is already available online (I had one knocked back from MEE for this reason recently).
