### 2018-06-13 #212 @sckott
@shaunpwilkinson good question about JOSS, i don't know. @noamross do you know answer to the above? whether JOSS cares if the paper is also somewhere else in another form?
