### 2018-06-18 #212 @shaunpwilkinson
Hi @sckott I can't see an invite to join a team within the ropensci organization - any chance you could point me in the direction of where I should be looking? Perhaps I'm just being blind.
