### 2018-06-19 #212 @shaunpwilkinson
No worries, almost there, I've managed to join the organization but when I go to transfer the repo I get the error message `You don’t have the permission to create repositories on ropensci`
