### 2018-06-20 #212 @sckott
woops, sorry. Github recently changed how permissions work for transferring repos to organizations.

can you add me as admin to your repo - then i'll transfer to ropensci - then you can take it from there?
