### 2018-06-20 #212 @sckott
thanks! not quite.

i need to be able to get to Settings tab, so i need admin access
