### 2018-06-20 #212 @shaunpwilkinson
Hmm, I don't seem to have that option, I just had a look at the GitHub help pages  (https://help.github.com/articles/permission-levels-for-a-user-account-repository/#collaborator-access-on-a-repository-owned-by-a-user-account) and it looks like there are only two permission levels, owner and collaborator, so I have just transferred the ownership to you
