### 2018-06-25 #212 @shaunpwilkinson
Hi @sckott and @stefaniebutland, yes I'd be more than happy to, I'll write up a technote asap.

Also just FYI the latest version of the package is on CRAN now, and JOSS paper is online too (http://joss.theoj.org/papers/10.21105/joss.00790)
