### 2018-06-26 #212 @stefaniebutland
Glad to hear you'll contribute a technote @shaunpwilkinson. Will tweet it from @rOpenSci to get more eyes on your work.

Here's info on how to submit via pull request https://github.com/ropensci/roweb2#contributing-a-blog-post. Difference in YAML:
> categories: technote

Since Tech Notes can be published any time, you can choose the date, but please submit about a week in advance so we can review. Happy to answer any questions that come up.
