### 2018-06-26 #212 @stefaniebutland
Forgot to mention, please add these tags to YAML (among others)
- onboarding
- review
- r

And be sure that the tech note gives an example of a cool thing you can do with the package, rather than reiterating the README for example.

Thank you!
